fix: tests and moved course functions to lms utils
