+++ school/lms/widgets/CourseCard.html
@@ -31,7 +31,7 @@
       <span class="sr-only"> {{ progress }} Complete</span>
     </div>
   </div>
-  <div class="progress-percentage">{{ progress }}% Completed</div>
+  <div class="progress-percent">{{ progress }}% Completed</div>
   {% endif %}
 
   <div class="zindex course-card-footer">
